Add FireBurst

+++ bending-paper/src/main/resources/mariadb.sql
@@ -31,7 +31,7 @@ CREATE INDEX IF NOT EXISTS elements_index ON bending_players_elements (player_id
 
 CREATE TABLE IF NOT EXISTS bending_players_slots (
     player_id       INT                     NOT NULL,
-    slot            SMALLINT                NOT NULL,
+    slot            INT                     NOT NULL,
     ability_id      INT                     NOT NULL,
     FOREIGN KEY(player_id) REFERENCES bending_players(player_id) ON DELETE CASCADE,
     FOREIGN KEY(ability_id) REFERENCES bending_abilities(ability_id) ON DELETE CASCADE,
@@ -49,7 +49,7 @@ CREATE TABLE IF NOT EXISTS bending_presets (
 
 CREATE TABLE IF NOT EXISTS bending_presets_slots (
     preset_id       INT                     NOT NULL,
-    slot            SMALLINT                NOT NULL,
+    slot            INT                     NOT NULL,
     ability_id      INT                     NOT NULL,
     FOREIGN KEY(preset_id) REFERENCES bending_presets(preset_id) ON DELETE CASCADE,
     FOREIGN KEY(ability_id) REFERENCES bending_abilities(ability_id) ON DELETE CASCADE,
