+++ bending-paper/src/main/resources/sqlite.sql
@@ -28,7 +28,7 @@ CREATE INDEX elements_index ON bending_players_elements (player_id);
 
 CREATE TABLE IF NOT EXISTS bending_players_slots (
     player_id       INTEGER                 NOT NULL,
-    slot            SMALLINT                NOT NULL,
+    slot            INTEGER                 NOT NULL,
     ability_id      INTEGER                 NOT NULL,
     FOREIGN KEY(player_id) REFERENCES bending_players(player_id) ON DELETE CASCADE,
     FOREIGN KEY(ability_id) REFERENCES bending_abilities(ability_id) ON DELETE CASCADE,
@@ -45,7 +45,7 @@ CREATE TABLE IF NOT EXISTS bending_presets (
 
 CREATE TABLE IF NOT EXISTS bending_presets_slots (
     preset_id       INTEGER                 NOT NULL,
-    slot            SMALLINT                NOT NULL,
+    slot            INTEGER                 NOT NULL,
     ability_id      INTEGER                 NOT NULL,
     FOREIGN KEY(preset_id) REFERENCES bending_presets(preset_id) ON DELETE CASCADE,
     FOREIGN KEY(ability_id) REFERENCES bending_abilities(ability_id) ON DELETE CASCADE,
